restructured UpdateData for modular fie structure
